+++ previous-versions/commerce/RELEASES.md
@@ -617,3 +617,13 @@ Based on:
 - [python v0.1.0] previous-versions/commerce
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-commerce/0.1.0 - previous-versions/commerce
+
+## 2023-09-07 00:18:01
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Commerce.yaml
+- Speakeasy CLI 1.79.0 (2.101.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0] previous-versions/commerce
+### Releases
+- [PyPI v0.1.0] https://pypi.org/project/codat-commerce/0.1.0 - previous-versions/commerce